requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
numpy==1.24.3
opencv-python==4.7.0.72
torch==1.12.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
ultralytics==8.0.124
xgboost==1.7.5
pycryptodome==3.17
python-multipart==0.0.6